backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
neomodel==5.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
asgiref==3.9.1
neo4j==5.15.0
pytz==2025.2
ruff==0.12.9
sqlparse==0.5.3